+++ src/examples/compose/app.ts
@@ -2,7 +2,7 @@ import { interfaceOf, Interfaces } from '../../core'
 import { stateOf } from '../../component'
 import { StyleGroup } from '../../style'
 import { View } from '../../interfaces/view'
-import h from 'snabbdom/h'
+import h from '../../interfaces/h'
 
 import * as Counter from './counter'