+++ src/examples/compose/counter.ts
@@ -2,7 +2,7 @@ import { Inputs, Actions, Interfaces, ev } from '../../core'
 import { StyleGroup, clickable } from '../../style'
 
 import { View } from '../../interfaces/view'
-import h from 'snabbdom/h'
+import h from '../../interfaces/h'
 
 export const name = 'Counter'