+++ src/examples/router/sideMenu.ts
@@ -1,7 +1,7 @@
 import { Actions, Inputs, Interfaces } from '../../core'
 import { StyleGroup, clickable } from '../../style'
 import { View } from '../../interfaces/view'
-import h from 'snabbdom/h'
+import h from '../../interfaces/h'
 import { palette } from './constants'
 
 export const name = 'Root'