+++ src/examples/worker/counter.ts
@@ -1,8 +1,7 @@
 import { Component, ev, Inputs } from '../../core'
 import { StyleGroup, clickable } from '../../style'
-
 import { View } from '../../interfaces/view'
-import h from 'snabbdom/h'
+import h from '../../interfaces/h'
 
 let name = 'Counter'